python/requirements.txt
pyaudio 
vosk 
sounddevice 
numpy 
piper-tts>1.3.0
pyusb 
requests
onnxruntime 
webrtcvad
setuptools<81